+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 ### Added
 - Move pipeline dereference methods into utils file
+- Getter to retrieve available pipelines in steps form
 
 ### Changed
 - A pipeline can start by a reference to an other pipeline